Completed Core Ability System
you can created abilities that perform different actions as well as setting them as chargeable so if you hold the button it stalls you midair while you cast as well as options that allow you to move while casting. It also detects hitting others that have a damagable tag. Should work on projectile to see how that goes.