ensured that level up and ui works for all clients
it was kinda weird bc i didnt know that if a nav mesh has a rigid body it NEEDS to be kinematic but thats a thing. so now it doesn't do the weird bounce in floor thing. so now pretty much everything about fights work in multiplayer its kinda wilding