adding charging particles and returned to piece by piece prefab
basically i had made the player combatant gameobject a rendered texture bc my dumbahh thought that sprites in a 3d environment cause issues with render order but no im just dumb. so i returned it back to how it was and readjusted everything. ALSO added in particles that occur when you're in startup/charging an attack (so hopefully later it'll indicate someone is about to attack)