w0w
learned a LOT about shaders to make different effects including a nice glow effect for magic that i liked a lot. I also realized that in order for charging to work i HAVE to seperate the player animators attack state into start up and release. But that freedom allows for many more shinanigans with animations hopefully