finished dropping items when full
basically before it just wouldnt give you an item if you were full but NOW it gives you the choice to drop an item if that inventory is full, which is the tricky part since there's 3 different inventories. SWAG. honestly its a pain to do stuff like this bc there's no logic or pizzazz just a lot of UI and syncing. still gotta pvp victory stuff tho